sha1.c und aes.c renamed
