+++ C_SDK/crypt/cubesql_sha1.h
